[fix]: denomination keyword replaced with major/minor
