+++ .changeset/funny-dingos-worry.md
@@ -1,5 +1,5 @@
 ---
-"@razorpay/i18nify-js": patch
+'@razorpay/i18nify-js': patch
 ---
 
-[feat]: add denomination conversion apis in currency module
+[feat]: add major/minor conversion apis in currency module
