+++ packages/i18nify-js/src/modules/currency/__tests__/convertToMajorUnit.test.ts
@@ -1,19 +1,19 @@
-import convertToLowerDenomination from '../convertToLowerDenomination';
+import convertToMajorUnit from '../convertToMajorUnit';
 import { CURRENCIES } from '../data/currencies';
 
-describe('currency - convertToLowerDenomination', () => {
+describe('currency - convertToMajorUnit', () => {
   const testCases: {
     amount: number;
     currency: keyof typeof CURRENCIES;
     expectedResult: number;
   }[] = [
-    { amount: 1, currency: 'USD', expectedResult: 100 },
-    { amount: 1, currency: 'GBP', expectedResult: 100 },
+    { amount: 100, currency: 'USD', expectedResult: 1 },
+    { amount: 100, currency: 'GBP', expectedResult: 1 },
   ];
 
   testCases.forEach(({ amount, currency, expectedResult }) => {
-    it(`should correctly convert ${amount} of lower denomination ${currency} to ${expectedResult}`, () => {
-      const result = convertToLowerDenomination(amount, { currency: currency });
+    it(`should correctly convert ${amount} of minor unit ${currency} to ${expectedResult}`, () => {
+      const result = convertToMajorUnit(amount, { currency: currency });
       expect(result).toBe(expectedResult);
     });
   });
@@ -22,7 +22,7 @@ describe('currency - convertToLowerDenomination', () => {
     const unsupportedCurrencyCode = 'XXX';
     expect(() => {
       // @ts-expect-error intented invalid currencyCode for testing
-      convertToLowerDenomination(100, { currency: unsupportedCurrencyCode });
+      convertToMajorUnit(100, { currency: unsupportedCurrencyCode });
     }).toThrow('Unsupported currency XXX');
   });
 });
