+++ packages/i18nify-js/src/modules/currency/__tests__/convertToMinorUnit.test.ts
@@ -1,19 +1,19 @@
-import convertToUpperDenomination from '../convertToUpperDenomination';
+import convertToMinorUnit from '../convertToMinorUnit';
 import { CURRENCIES } from '../data/currencies';
 
-describe('currency - convertToUpperDenomination', () => {
+describe('currency - convertToMinorUnit', () => {
   const testCases: {
     amount: number;
     currency: keyof typeof CURRENCIES;
     expectedResult: number;
   }[] = [
-    { amount: 100, currency: 'USD', expectedResult: 1 },
-    { amount: 100, currency: 'GBP', expectedResult: 1 },
+    { amount: 1, currency: 'USD', expectedResult: 100 },
+    { amount: 1, currency: 'GBP', expectedResult: 100 },
   ];
 
   testCases.forEach(({ amount, currency, expectedResult }) => {
-    it(`should correctly convert ${amount} of lower denomination ${currency} to ${expectedResult}`, () => {
-      const result = convertToUpperDenomination(amount, { currency: currency });
+    it(`should correctly convert ${amount} of minor unit ${currency} to ${expectedResult}`, () => {
+      const result = convertToMinorUnit(amount, { currency: currency });
       expect(result).toBe(expectedResult);
     });
   });
@@ -22,7 +22,7 @@ describe('currency - convertToUpperDenomination', () => {
     const unsupportedCurrencyCode = 'XXX';
     expect(() => {
       // @ts-expect-error intented invalid currencyCode for testing
-      convertToUpperDenomination(100, { currency: unsupportedCurrencyCode });
+      convertToMinorUnit(100, { currency: unsupportedCurrencyCode });
     }).toThrow('Unsupported currency XXX');
   });
 });
