+++ packages/i18nify-js/src/modules/currency/index.ts
@@ -2,5 +2,5 @@ export { default as formatNumber } from './formatNumber';
 export { default as getCurrencyList } from './getCurrencyList';
 export { default as getCurrencySymbol } from './getCurrencySymbol';
 export { default as formatNumberByParts } from './formatNumberByParts';
-export { default as convertToUpperDenomination } from './convertToUpperDenomination';
-export { default as convertToLowerDenomination } from './convertToLowerDenomination';
+export { default as convertToMajorUnit } from './convertToMajorUnit';
+export { default as convertToMinorUnit } from './convertToMinorUnit';
